--- the-news/src/components/API/CodingArticles.js
@@ -1,10 +1,0 @@
-import axios from 'axios';
-                
-export const fetchCodingArticles = () => {
-    return axios
-    .get("https://nc-news-example-1.herokuapp.com/api/articles?topic=coding")
-    .then(({data}) => {
-        return data.articles
-    });
-}
-
