+++ the-news/src/components/API/GetArticleByTopic.js
@@ -1,0 +1,9 @@
+import axios from 'axios';
+
+export const fetchArticleByTopic = (topic) => {
+    return axios
+    .get(`https://nc-news-example-1.herokuapp.com/api/articles?topic=${topic}`)
+    .then(({data}) => {
+        return data.articles
+    });
+}
