+++ the-news/src/components/Routes/TopicsList.js
@@ -1,6 +1,7 @@
-import fetchTopics from './API/Topics';
+import fetchTopics from '../API/Topics';
 import {useState, useEffect} from 'react';
 import { Link } from 'react-router-dom';
+import ArticleByTopic from '../Routes/ArticleByTopic';
 
 
 export default function TopicList() {
@@ -12,15 +13,13 @@ export default function TopicList() {
             setTopics(topicsList)
         })
     }, [])
-    
-
 
     return (
         <ul className="topics-list">
             {topics.map((topic) => {
                 return (
                     <section key ={topic.slug} >
-                    <Link style={{color: 'black'}} className="topic-item"to={`/topics/${topic.slug}`}>{topic.slug}</Link>
+                    <Link onClick={<ArticleByTopic topic={topic} />} style={{color: 'black'}} className="topic-item"to={`/topics/${topic.slug}`}>{topic.slug}</Link>
                     </section>
                 )
             })}
